Upload a file via an external binary
Upload a file by dropping the file onto the nick selector.

The uploader binary needs to take the full path of the file
as its first argument and needs to echo ONLY the full url
where the file ended up, this will be pasted into the input.

The uploader binary location is configured via
quasselClient.conf, no UI at the moment

[Config]
UploadCommand=/path/to/uploader